requirements.txt
PyYAML>=6.0 #Config
requests>=2.31.0 #HTTP and API
openai>=1.17.1 #LLM